Added global AI chatbot component to all pages Created reusable chatbot component and callbacks for book recommendations. Chatbot persists conversation history across pages and personalizes responses
